requirements.txt
requests
lxml
bs4
python-dotenv